==> 13_mapmaking solutions.Rmd <==
---
title: "Module 3 Solutions"
author: "Brenna Kelly"
date: "2024-08-10"
output: 
  html_document:
    toc: true
    toc_float: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r message=FALSE, warning=FALSE}

library(sf)
library(tmap)
library(ggpubr)
library(ggplot2)
library(ggspatial)
library(prettymapr)

```

## `tmap` and `ggplot2`

There are two main R packages for plotting static maps: [tmap](https://r-tmap.github.io/tmap/) and [ggplot2](https://ggplot2.tidyverse.org/reference/ggsf.html) (R users may be more familiar with ggplot2). With both packages, you can visualize simple feature (sf) objects.

```{r}

guerry <- st_read("data/guerry/guerry.shp")

tm_shape(guerry) +
  tm_polygons()

ggplot(guerry) +
  geom_sf()

```

As you start creating your own maps, you may find you prefer one or the other, but each package has useful features.  

Notice, for instance, that we don't need to specify the type of geometry with ggplot, instead simply using `geom_sf` (unlike tmap, which requires tm_polygon, tm_dots, tm_lines).

```{r}

guerry_poly <- ggplot(guerry) +
  geom_sf()

guerry_ln <- st_cast(guerry, "MULTILINESTRING")
guerry_line <- ggplot(guerry_ln) +
  geom_sf()

guerry_pt <- st_centroid(guerry)
guerry_dot <- ggplot(guerry_pt) +
  geom_sf()

ggarrange(guerry_poly, guerry_line, guerry_dot, nrow = 1, ncol = 3)

```

We can also add labels using `geom_text` (or `geom_label`, if you want a box behind the text, for visibility).

```{r}

ggplot(guerry) +
  geom_sf() +
  geom_text(mapping = aes(label = Dprtmnt, geometry = geometry), 
            stat = "sf_coordinates", position = "identity", size = 4)

ggplot(guerry) +
  geom_sf() +
  geom_label(aes(label = Dprtmnt, geometry = geometry), 
            stat = "sf_coordinates", position = "identity", size = 3)

```

You can also add basemaps with tiles from the ggspatial packages.  

```{r}

ggplot(guerry) +
  annotation_map_tile(type = "osm", cachedir = "rosm.cache/", zoomin = -1) +
  geom_sf(aes(fill = Pop1831), lwd = 0) +
  scale_fill_viridis_c(option = "D")

```

The tmap package has many of the same options, but with slightly different syntax. As you explore your data, you may also like the dynamic `view` feature, which uses [leaflet](https://rstudio.github.io/leaflet/index.html), a popular package for interactive maps:

```{r}

house <- st_read("data/clean_data/house.shp")
pumps <- st_read("data/clean_data/pumps.shp")

tmap_mode("view")
tm_shape(house) +
  tm_dots(col = "name", style = "cont", 
          palette = "Dark2", legend.show = FALSE) +
  tm_shape(pumps) +
  tm_dots(col = "name", size = 0.5, title = "Water pump",
          palette = "Dark2", shape = 17) +
  tm_layout(title = "Nearest water pump to\neach household", main.title.position = "CENTER")

tmap_mode("plot")
tm_shape(house) +
  tm_dots(col = "name", style = "cont", 
          palette = "Dark2", legend.show = FALSE) +
  tm_shape(pumps) +
  tm_dots(col = "name", size = 1, title = "Water pump",
          palette = "Dark2", shape = 17) +
  tm_layout(title = "Nearest water pump to\neach household", main.title.position = "CENTER")

```

## Mapping populations  
As population health scientists, we are often interested in mapping populations. A common approach would be to aggregate the number of cases by geographic unit (e.g., county, hospital referral region) and plot these areal units as polygons:  
```{r}

ca <- st_read("data/california/ca_counties.shp")

tm_shape(ca) +
  tm_polygons(col = "f_ra", palette = "RdPu", style = "cont", lwd = 0,
              title = "Females of \nreproductive age")

tm_shape(ca) +
  tm_polygons(col = "f_ra_p", palette = "RdPu", style = "cont", lwd = 0,
              title = "% of \nreproductive age")

```

However, that approach shows us a map of places, not people. A better method would visually represent the areas based on the populations they contain.  

The Census holds geographic coordinates for [centers of population](https://www.census.gov/geographies/reference-files/time-series/geo/centers-population.html), a really fantastic resource I highly recommend. To create an sf object from these csvs, you would follow similar procedures to what we did in `01_spatial data.R` with the spirals dataset.

When we represent counties with population-weighted centroids scaled by the population size, we get a more accurate representation:  

```{r}

ca_centers <- st_read("data/california/ca_county_centers.shp")

tm_shape(ca) +
  tm_polygons(col = "white", lwd = 0.25, title = "") +
  tm_shape(ca_centers) +
  tm_dots(col = "f_ra_p", size = "f_ra", scale = 2,
          palette = "RdPu", style = "cont", legend.is.portrait = FALSE,
          title = "% of Reproductive Age")

```

Create similar maps to what we saw above with California, but for census tracts in Los Angeles County. Compare the polygon and dot maps. How does your interpretation change between maps?

```{r}

la <- st_read("data/california/la_tracts.shp")

la_centers <- st_read("data/california/la_centers.shp")

tm_shape(la) +
  tm_polygons(col = "f_ra_p", palette = "RdPu", style = "cont", lwd = 0, 
              legend.is.portrait = TRUE, title = "Proportion \nof Reproductive Age")

tm_shape(la) +
  tm_polygons(col = "white", lwd = 0.25) +
  tm_shape(la_centers) +
  tm_dots(col = "f_ra_p", style = "cont", palette = "RdPu", 
          size = "f_ra", scale = 0.5, alpha = 0.7,
          legend.is.portrait = FALSE, title = "Proportion \nof Reproductive Age")

```
